+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.41
+
+Use the same JWT library instead of using two different ones.
+
+Add Description() for Alert bodies.
+
+## 0.40
+
+Fix next page URL's for Twilio Monitor
+
 ## 0.39
 
 The data in Update() requests was silently being ignored. They are not ignored